Added server creation status and createServer method
In the servers.component.ts file, a new variable 'serverCreationStatus' was added to keep track of the server creation status. A new method 'createServer' was also added that changes the value of 'serverCreationStatus' to indicate a new server has been created. These changes were reflected in the servers.component.html file where a click event was added to the 'Add a new server' button to trigger the 'createServer' method and the 'serverCreationStatus' is now displayed on the UI. This was done to provide server creation feedback to the user.