Add server name input field and dynamic update
A new input field and an event listener 'input' was added in servers.component.html. This binds to method 'updateServerName' which dynamically updates 'serverName' value in servers.component.ts. Moreover, server name display logic is added to show the current server name or a request to provide one. This change allows users to input and see the server name in real time.